[Refact]mainPage안에 있는 컴포넌트 App.js로 이동
